+++ src/prod.js
@@ -4,6 +4,9 @@ import * as _hooks from 'preact/hooks';
 import _useResource from './hooks/useResource/index.js';
 import _useUid from './hooks/useUid.js';
 import _useScheduled from './hooks/useScheduled.js';
+import _useEventListener from './hooks/useEventListener.js';
+import _useSharedState from './hooks/useSharedState.js';
+
 import _createI18nContext from './createI18nContext.js';
 import _asyncComponents from './asyncComponents.js';
 
@@ -21,7 +24,10 @@ const _inject = { ..._hooks, createElement: h, createContext };
 export const useResource = _useResource(_inject);
 export const useUid = _useUid(_inject);
 export const useScheduled = _useScheduled(_inject);
+export const useEventListener = _useEventListener(_inject);
+export const useSharedState = _useSharedState(_inject);
+
 export const { asyncComponent, asyncComponents } = _asyncComponents(_inject);
 export const createI18nContext = _createI18nContext(_inject);
 
-export const VERSION = '1.3.1';
+export const VERSION = '1.3.3';